payment/requirements.txt
azure-functions
mysql-connector-python==8.2.0
